move selectors to selectors
