+++ frontend/react/src/store/selectors.js
@@ -1,5 +1,41 @@
 import jsonpath from "jsonpath";
 
+import { selectFragment } from "./formData";
+
+export const selectSectionTitle = (state, sectionId) => {
+  const jspath = `$..formData[*].contents.section[?(@.id=='${sectionId}')].title`;
+  const sectionTitles = jsonpath.query(state, jspath);
+
+  if (sectionTitles.length) {
+    return sectionTitles[0];
+  }
+  return null;
+};
+
+export const selectSubsectionTitleAndPartIDs = (state, subsectionId) => {
+  const subsection = selectFragment(state, subsectionId);
+
+  if (subsection) {
+    return {
+      parts: subsection.parts.map((part) => part.id),
+      title: subsection.title,
+    };
+  }
+  return null;
+};
+
+export const selectPartTitle = (state, partId) => {
+  const part = selectFragment(state, partId);
+
+  if (part) {
+    return {
+      text: part.text,
+      title: part.title,
+    };
+  }
+  return null;
+};
+
 export const selectQuestion = (state, id) => {
   const jp = `$..[*].contents.section.subsections[*].parts[*]..questions[?(@.id=='${id}')]`;
   const questions = jsonpath.query(state, jp);
